# docs/openapi.yaml
openapi: 3.1.0
info:
  title: EV Battery Telemetry & Diagnostics
  description: Ingest, query, and analyse EV battery telemetry data.
  version: 1.0.0
paths:
...
  /api/v1/rul/{battery_id}:
    get:
      tags:
      - RUL Predictions
      summary: Get latest RUL prediction details for a battery
      description: 'Return the latest computed RUL prediction, confidence interval,
        and alert classification.

        Available to authenticated fleet_admin and operator users.'
      operationId: get_rul_api_v1_rul__battery_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: battery_id
        in: path
        required: true
        schema:
          type: string
          title: Battery Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RULResponse'
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ConfidenceInterval:
      properties:
        lower_bound:
          type: integer
          title: Lower Bound
          description: Lower interval bound of RUL prediction
          examples:
          - 188
        upper_bound:
          type: integer
          title: Upper Bound
          description: Upper interval bound of RUL prediction
          examples:
          - 238
        confidence_percent:
          type: number
          title: Confidence Percent
          description: Statistical confidence level percent
          default: 90.0
          examples:
          - 90.0
      type: object
      required:
      - lower_bound
      - upper_bound
      title: ConfidenceInterval
...
    ErrorResponse:
      properties:
        error:
          type: string
          title: Error
          description: Error description message
          examples:
          - Rate limit exceeded
      type: object
      required:
      - error
      title: ErrorResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RULResponse:
      properties:
        battery_id:
          type: string
          title: Battery Id
          description: Battery pack ID
          examples:
          - B0047
        predicted_rul_cycles:
          type: integer
          title: Predicted Rul Cycles
          description: LSTM predicted remaining cycles before EOL threshold is hit
          examples:
          - 213
        confidence_interval:
          $ref: '#/components/schemas/ConfidenceInterval'
        current_soh_percent:
          type: number
          title: Current Soh Percent
          description: SOH percentage used for this prediction
          examples:
          - 83.71
        eol_threshold_soh:
          type: number
          title: Eol Threshold Soh
          description: SOH target threshold representing End Of Life
          default: 70.0
          examples:
          - 70.0
        model_version:
          type: string
          title: Model Version
          description: Model definition version code
          examples:
          - v2.0
        predicted_at:
          type: string
          format: date-time
          title: Predicted At
          description: Timestamp of prediction generation
          examples:
          - '2026-06-13T17:03:00Z'
        alert_level:
          type: string
          title: Alert Level
          description: Alert classification ('healthy', 'warning', or 'critical')
          examples:
          - healthy
      type: object
      required:
      - battery_id
      - predicted_rul_cycles
      - confidence_interval
      - current_soh_percent
      - model_version
      - predicted_at
      - alert_level
      title: RULResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /auth/token